backend-python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
snowflake-connector-python>=3.12.0
python-dotenv==1.0.0
requests>=2.31.0
tabulate>=0.9.0
